Fixes issue with duplicate keys not being thrown
It was reported that if duplicate keys are encountered during read, no
exception was thrown and the read would abruptly end without notice.
This was caused by a try/catch within the Read method. There was an
exception being thrown when adding the key value but was being caught by
the Read method. This commit removes that try/catch so that it can be
handled the caller.
This commit also refactors the Read method to remove all the if/else
nesting and clean up the code a bit.

fixes #3